Moved csrf context from components to contexts
